clarify that rewrapping is tied to auth
